FIX: Тип поля "Сумма заказа" у структуры "Order" модуля "Shop" изменён на "float"

+++ Shop/Structure/Order/config.php
@@ -33,7 +33,7 @@ return array(
         ),
         'price' => array(
             'label' => 'Сумма заказа',
-            'sql'   => 'int',
+            'sql'   => 'float',
             'type'  => 'Shop_Price'
         ),
         'stock' => array(
